requirements.txt
fpdf2>=2.7.0
fastapi>=0.115.0
uvicorn>=0.30.0
streamlit>=1.38.0
httpx>=0.27.0
python-dotenv>=1.0.1
pydantic>=2.8.0
plotly>=5.24.0
duckduckgo-search>=6.2.13
langchain>=0.2.14
langchain-community>=0.2.12
langchain-openai>=0.1.22
langchain-ollama>=0.1.3
faiss-cpu>=1.8.0.post1
pypdf>=4.3.1
